+++ src/main/java/io/macgyver/reactor/aws/sns/SNSAdapter.java
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import io.macgyver.reactor.aws.sqs.SQSReactorBridge.SQSMessage;
+import io.macgyver.reactor.aws.sqs.SQSMessage;
 import io.macgyver.reactor.aws.util.MoreSelectors;
 import reactor.bus.Event;
 import reactor.bus.EventBus;